requirements.txt
addict==2.4.0
backcall==0.2.0
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
configparser==5.2.0
contextlib2==21.6.0
cycler==0.11.0
dataclasses
decorator==4.4.2
dill==0.3.4
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.18
h5py==3.1.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.8.3
ipdb==0.13.7
ipython==7.16.2
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mpi4py==3.0.3
mpmath==1.2.1
multiprocess==0.70.11.1
networkx==2.5.1
numpy==1.19.2
olefile==0.46
opencv-python==4.4.0.46
packaging==21.3
paho-mqtt==1.5.1
pandas==1.1.4
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2021.11.10
requests==2.27.1
rsa==4.6
sacremoses==0.0.43
schema==0.7.4
scikit-image==0.17.2
scikit-learn==0.22.1
scipy==1.5.4
seaborn==0.11.2
sentry-sdk==1.5.1
setproctitle==1.1.10
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
sympy==1.9
termcolor==1.1.0
tifffile==2020.9.3
timm==0.4.12
toml==0.10.2
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.49.0
traitlets==4.3.3
typing-extensions
wandb==0.12.9
watchdog==2.1.6
wcwidth==0.2.5
yacs==0.1.8
yapf==0.32.0
yaspin==2.1.0
zipp==3.6.0
